pyproject.toml
[project]
name = "content_creator"
version = "0.1.0"
description = "content creator using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.175.0,<1.0.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
content_creator = "content_creator.main:run"
run_crew = "content_creator.main:run"
train = "content_creator.main:train"
replay = "content_creator.main:replay"
test = "content_creator.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
